Fix warnings

+++ src/Internal/Debug.cpp
@@ -5,7 +5,7 @@
 
 void Debug::Print(const char* str)
 {
-	for (int i = 0; i < strlen(str); i++) SerialUSB.print(str[i]);
+	for (int i = 0; (unsigned)i < strlen(str); i++) SerialUSB.print(str[i]);
 }
 
 void Debug::Println (const char* str)
